UPDATE auth integration testing
mod: restructured auth routers
