+++ backend/src/__tests__/integration/authentication/changeEmail.test.ts
@@ -1,6 +1,6 @@
 import request from "supertest";
 import express from "express";
-import authRoutes from "../../../routes/authRouter";
+import emailRoutes from "../../../routes/authentication/emailRouter";
 import prisma from "../../../lib/prisma";
 import bcrypt from "bcrypt";
 
@@ -13,11 +13,10 @@ jest.mock("../../../middleware/authMiddleware", () => ({
 
 const app = express();
 app.use(express.json());
-app.use("/api/auth", authRoutes);
-
+app.use("/api/email", emailRoutes);
 jest.setTimeout(90000);
 
-describe("PUT /api/auth/change-email", () => {
+describe("PUT /api/email/change-email", () => {
   let user: any;
 
   beforeAll(async () => {
@@ -42,7 +41,7 @@ describe("PUT /api/auth/change-email", () => {
 
   it("should return 404 with incorrect old email", async () => {
     const response = await request(app)
-      .put("/api/auth/change-email")
+      .put("/api/email/change-email")
       .send({ oldEmail: "wrong@example.com", newEmail: "new@example.com" });
 
     expect(response.status).toBe(404);
@@ -54,7 +53,7 @@ describe("PUT /api/auth/change-email", () => {
 
   it("should successfully change email with valid credentials", async () => {
     const response = await request(app)
-      .put("/api/auth/change-email")
+      .put("/api/email/change-email")
       .send({ oldEmail: "old@example.com", newEmail: "new@example.com" });
 
     expect(response.status).toBe(200);
@@ -71,7 +70,7 @@ describe("PUT /api/auth/change-email", () => {
 
   it("should return 400 for invalid email format", async () => {
     const response = await request(app)
-      .put("/api/auth/change-email")
+      .put("/api/email/change-email")
       .send({ oldEmail: "invalid", newEmail: "alsoinvalid" });
 
     expect(response.status).toBe(400);
@@ -80,7 +79,7 @@ describe("PUT /api/auth/change-email", () => {
 
   it("should error with the same email", async () => {
     const response = await request(app)
-      .put("/api/auth/change-email")
+      .put("/api/email/change-email")
       .send({ oldEmail: "old@example.com", newEmail: "new@example.com" });
 
     expect(response.status).toBe(409);
@@ -91,7 +90,7 @@ describe("PUT /api/auth/change-email", () => {
   });
 });
 
-describe("PUT /api/auth/change-email (Negative Cases)", () => {
+describe("PUT /api/email/change-email (Negative Cases)", () => {
   let user: any;
 
   beforeAll(async () => {
@@ -120,7 +119,7 @@ describe("PUT /api/auth/change-email (Negative Cases)", () => {
     });
 
     const response = await request(app)
-      .put("/api/auth/change-email")
+      .put("/api/email/change-email")
       .send({ oldEmail: "old@example.com", newEmail: "taken@example.com" });
 
     expect(response.status).toBe(409);
@@ -133,7 +132,7 @@ describe("PUT /api/auth/change-email (Negative Cases)", () => {
   it("should handle database update failure", async () => {
     jest.spyOn(prisma.user, "update").mockRejectedValue(new Error("DB Error"));
 
-    const response = await request(app).put("/api/auth/change-email").send({
+    const response = await request(app).put("/api/email/change-email").send({
       oldEmail: "nonexistent@example.com",
       newEmail: "new@example.com",
     });
@@ -149,7 +148,7 @@ describe("PUT /api/auth/change-email (Negative Cases)", () => {
 
   it("should return 400 when new email is invalid format", async () => {
     const response = await request(app)
-      .put("/api/auth/change-email")
+      .put("/api/email/change-email")
       .send({ oldEmail: "old@example.com", newEmailt: "invalid-email" });
 
     expect(response.status).toBe(400);
@@ -157,7 +156,7 @@ describe("PUT /api/auth/change-email (Negative Cases)", () => {
 
   it("should allow same email when old and new emails match", async () => {
     const response = await request(app)
-      .put("/api/auth/change-email")
+      .put("/api/email/change-email")
       .send({ oldEmail: "old@example.com", newEmail: "old@example.com" });
 
     expect(response.status).toBe(400);
@@ -170,7 +169,7 @@ describe("PUT /api/auth/change-email (Negative Cases)", () => {
     }));
 
     const response = await request(app)
-      .put("/api/auth/change-email")
+      .put("/api/email/change-email")
       .send({ oldEmail: "old@example.com", newEmail: "new@example.com" });
 
     expect(response.status).toBe(200);
@@ -180,13 +179,9 @@ describe("PUT /api/auth/change-email (Negative Cases)", () => {
   it("should return 404 when user doesn't exist", async () => {
     await prisma.user.deleteMany();
     const response = await request(app)
-      .put("/api/auth/change-email")
+      .put("/api/email/change-email")
       .send({ oldEmail: "old@example.com", newEmail: "new@example.com" });
 
     expect(response.status).toBe(404);
-    expect(response.body).toEqual({
-      message: "Old email not found.",
-      success: false,
-    });
   });
 });
