+++ backend/src/__tests__/integration/authentication/login.test.ts
@@ -1,13 +1,13 @@
 import request from "supertest";
 import express from "express";
-import authRoutes from "../../../routes/authRouter";
+import userRoutes from "../../../routes/authentication/userRouter";
 import prisma from "../../../lib/prisma";
 import bcrypt from "bcrypt";
 import jwt from "jsonwebtoken";
 
 const app = express();
 app.use(express.json());
-app.use("/api/auth", authRoutes);
+app.use("/api/user", userRoutes);
 
 jest.setTimeout(60000);
 
@@ -15,7 +15,7 @@ jest.mock("jsonwebtoken", () => ({
   sign: jest.fn(() => "mocked-jwt-token"),
 }));
 
-describe("POST /api/auth/login (Negative Cases)", () => {
+describe("POST /api/user/login (Negative Cases)", () => {
   beforeAll(async () => {
     await prisma.otp.deleteMany();
     await prisma.user.deleteMany();
@@ -41,7 +41,7 @@ describe("POST /api/auth/login (Negative Cases)", () => {
       .mockRejectedValue(new Error("DB Failure"));
 
     const response = await request(app)
-      .post("/api/auth/login")
+      .post("/api/user/login")
       .send({ pin: "654321" });
 
     expect(response.status).toBe(500);
@@ -52,7 +52,7 @@ describe("POST /api/auth/login (Negative Cases)", () => {
 
   it("should return a token for valid PIN", async () => {
     const response = await request(app)
-      .post("/api/auth/login")
+      .post("/api/user/login")
       .send({ pin: "654321" });
 
     expect(response.status).toBe(200);
@@ -69,7 +69,7 @@ describe("POST /api/auth/login (Negative Cases)", () => {
       },
     });
     const response = await request(app)
-      .post("/api/auth/login")
+      .post("/api/user/login")
       .send({ pin: "121212" });
 
     expect(response.status).toBe(401);
@@ -77,14 +77,14 @@ describe("POST /api/auth/login (Negative Cases)", () => {
   });
 
   it("should return 400 if no PIN is provided", async () => {
-    const response = await request(app).post("/api/auth/login").send({});
+    const response = await request(app).post("/api/user/login").send({});
     expect(response.status).toBe(400);
     expect(response.body).toHaveProperty("message", "PIN cannot be empty");
   });
 
   it("should return 400 if PIN is too short", async () => {
     const response = await request(app)
-      .post("/api/auth/login")
+      .post("/api/user/login")
       .send({ pin: "12" });
 
     expect(response.status).toBe(400);
@@ -93,7 +93,7 @@ describe("POST /api/auth/login (Negative Cases)", () => {
 
   it("should return 400 if PIN contains invalid characters", async () => {
     const response = await request(app)
-      .post("/api/auth/login")
+      .post("/api/user/login")
       .send({ pin: "abc@12" });
 
     expect(response.status).toBe(400);
@@ -107,7 +107,7 @@ describe("POST /api/auth/login (Negative Cases)", () => {
     await prisma.user.deleteMany();
 
     const response = await request(app)
-      .post("/api/auth/login")
+      .post("/api/user/login")
       .send({ pin: "654321" });
 
     expect(response.status).toBe(404);
@@ -116,7 +116,7 @@ describe("POST /api/auth/login (Negative Cases)", () => {
 
   it("should return 400 for malformed request (not an object)", async () => {
     const response = await request(app)
-      .post("/api/auth/login")
+      .post("/api/user/login")
       .send("invalid body");
 
     expect(response.status).toBe(400);
@@ -125,7 +125,7 @@ describe("POST /api/auth/login (Negative Cases)", () => {
 
   it("should return 400 if PIN is null", async () => {
     const response = await request(app)
-      .post("/api/auth/login")
+      .post("/api/user/login")
       .send({ pin: null });
 
     expect(response.status).toBe(400);
@@ -133,7 +133,7 @@ describe("POST /api/auth/login (Negative Cases)", () => {
   });
 });
 
-describe("POST /api/auth/login (Negative Cases)", () => {
+describe("POST /api/user/login (Negative Cases)", () => {
   it("should handle bcrypt comparison failure after finding user", async () => {
     await prisma.user.create({
       data: {
@@ -147,7 +147,7 @@ describe("POST /api/auth/login (Negative Cases)", () => {
     });
 
     const response = await request(app)
-      .post("/api/auth/login")
+      .post("/api/user/login")
       .send({ pin: "654321" });
 
     expect([401, 500]).toContain(response.status);
@@ -156,7 +156,7 @@ describe("POST /api/auth/login (Negative Cases)", () => {
 
   it("should return 400 when PIN is a number instead of string", async () => {
     const response = await request(app)
-      .post("/api/auth/login")
+      .post("/api/user/login")
       .send({ pin: 123456 });
 
     expect(response.status).toBe(400);
@@ -164,7 +164,7 @@ describe("POST /api/auth/login (Negative Cases)", () => {
   });
 
   it("should return 400 when request body is empty", async () => {
-    const response = await request(app).post("/api/auth/login").send();
+    const response = await request(app).post("/api/user/login").send();
 
     expect(response.status).toBe(400);
     expect(response.body).toHaveProperty("message");
@@ -172,7 +172,7 @@ describe("POST /api/auth/login (Negative Cases)", () => {
 
   it("should handle JWT signing failure after successful auth", async () => {
     const normalResponse = await request(app)
-      .post("/api/auth/login")
+      .post("/api/user/login")
       .send({ pin: "654321" });
     expect(normalResponse.status).toBe(200);
 
@@ -181,7 +181,7 @@ describe("POST /api/auth/login (Negative Cases)", () => {
     });
 
     const response = await request(app)
-      .post("/api/auth/login")
+      .post("/api/user/login")
       .send({ pin: "654321" });
 
     expect([200, 500]).toContain(response.status);
