+++ backend/src/__tests__/integration/authentication/verifyEmail.test.ts
@@ -1,16 +1,16 @@
 import request from "supertest";
 import express from "express";
-import authRoutes from "../../../routes/authRouter";
+import emailRoutes from "../../../routes/authentication/emailRouter";
 import prisma from "../../../lib/prisma";
 import bcrypt from "bcrypt";
 
 const app = express();
 app.use(express.json());
-app.use("/api/auth", authRoutes);
+app.use("/api/email", emailRoutes);
 
 jest.setTimeout(90000);
 
-describe("POST /api/auth/verify-email", () => {
+describe("POST /api/email/verify-email", () => {
   beforeAll(async () => {
     await prisma.otp.deleteMany();
     await prisma.user.deleteMany();
@@ -30,7 +30,7 @@ describe("POST /api/auth/verify-email", () => {
 
   it("should verify correct email", async () => {
     const response = await request(app)
-      .post("/api/auth/verify-email")
+      .post("/api/email/verify-email")
       .send({ email: "test@example.com" });
 
     expect(response.status).toBe(200);
@@ -42,7 +42,7 @@ describe("POST /api/auth/verify-email", () => {
 
   it("should return 400 for incorrect email", async () => {
     const response = await request(app)
-      .post("/api/auth/verify-email")
+      .post("/api/email/verify-email")
       .send({ email: "wrong@example.com" });
 
     expect(response.status).toBe(400);
@@ -51,7 +51,7 @@ describe("POST /api/auth/verify-email", () => {
 
   it("should return 400 for empty email", async () => {
     const response = await request(app)
-      .post("/api/auth/verify-email")
+      .post("/api/email/verify-email")
       .send({ email: "" });
 
     expect(response.status).toBe(400);
@@ -59,27 +59,32 @@ describe("POST /api/auth/verify-email", () => {
   });
 });
 
-describe("POST /api/auth/verify-email (Negative Cases)", () => {
+describe("POST /api/email/verify-email (Negative Cases)", () => {
   it("should return 500 when database query fails", async () => {
     jest
       .spyOn(prisma.user, "findFirst")
       .mockRejectedValue(new Error("DB Error"));
 
     const response = await request(app)
-      .post("/api/auth/verify-email")
+      .post("/api/email/verify-email")
       .send({ email: "test@example.com" });
 
     expect(response.status).toBe(500);
     expect(response.body).toEqual({
       message: "Failed to verify email",
       success: false,
+      error: {
+        message: "DB Error",
+      },
     });
 
     jest.restoreAllMocks();
   });
 
   it("should return 400 when email is missing", async () => {
-    const response = await request(app).post("/api/auth/verify-email").send({});
+    const response = await request(app)
+      .post("/api/email/verify-email")
+      .send({});
 
     expect(response.status).toBe(400);
     expect(response.body).toEqual({ message: "Invalid email address" });
@@ -87,7 +92,7 @@ describe("POST /api/auth/verify-email (Negative Cases)", () => {
 
   it("should return 400 when request body is malformed", async () => {
     const response = await request(app)
-      .post("/api/auth/verify-email")
+      .post("/api/email/verify-email")
       .send("invalid-json");
 
     expect(response.status).toBe(400);
