+++ backend/src/__tests__/integration/authentication/verifyPin.test.ts
@@ -1,16 +1,16 @@
 import request from "supertest";
 import express from "express";
-import authRoutes from "../../../routes/authRouter";
+import pinRoutes from "../../../routes/authentication/pinRouter";
 import prisma from "../../../lib/prisma";
 import bcrypt from "bcrypt";
 
 const app = express();
 app.use(express.json());
-app.use("/api/auth", authRoutes);
+app.use("/api/pin", pinRoutes);
 
 jest.setTimeout(90000);
 
-describe("POST /api/auth/verify-pin", () => {
+describe("POST /api/pin/verify-pin", () => {
   beforeAll(async () => {
     await prisma.otp.deleteMany();
     await prisma.user.deleteMany();
@@ -30,7 +30,7 @@ describe("POST /api/auth/verify-pin", () => {
 
   it("should verify correct PIN", async () => {
     const response = await request(app)
-      .post("/api/auth/verify-pin")
+      .post("/api/pin/verify-pin")
       .send({ pin: "123456" });
 
     expect(response.status).toBe(200);
@@ -42,7 +42,7 @@ describe("POST /api/auth/verify-pin", () => {
 
   it("should return 401 for incorrect PIN", async () => {
     const response = await request(app)
-      .post("/api/auth/verify-pin")
+      .post("/api/pin/verify-pin")
       .send({ pin: "654321" });
 
     expect(response.status).toBe(401);
@@ -50,12 +50,12 @@ describe("POST /api/auth/verify-pin", () => {
   });
 });
 
-describe("POST /api/auth/verify-pin (Negative Cases)", () => {
+describe("POST /api/pin/verify-pin (Negative Cases)", () => {
   it("should return 404 when user doesn't exist", async () => {
     await prisma.user.deleteMany();
 
     const response = await request(app)
-      .post("/api/auth/verify-pin")
+      .post("/api/pin/verify-pin")
       .send({ pin: "123456" });
 
     expect(response.status).toBe(404);
@@ -66,13 +66,13 @@ describe("POST /api/auth/verify-pin (Negative Cases)", () => {
   });
 
   it("should return appropriate status when PIN is missing", async () => {
-    const response = await request(app).post("/api/auth/verify-pin").send({});
+    const response = await request(app).post("/api/pin/verify-pin").send({});
     expect([400, 401, 404]).toContain(response.status);
   });
 
   it("should return 400 when request body is malformed", async () => {
     const response = await request(app)
-      .post("/api/auth/verify-pin")
+      .post("/api/pin/verify-pin")
       .send("invalid-json");
 
     expect(response.status).toBe(404);
